@@ -4,12 +4,12 @@ plugins {
 }
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
 
     defaultConfig {
         applicationId "kr.co.hanbit.base"
         minSdkVersion 16
-        targetSdkVersion 30
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
 
@@ -36,6 +36,10 @@ android {
 }
 
 dependencies {
+    def dependency_version = "1.3.1"
+    implementation "androidx.activity:activity-ktx:$dependency_version"
+    implementation "androidx.fragment:fragment-ktx:$dependency_version"
+
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
