@@ -1,11 +1,9 @@
 package kr.co.hanbit.activity
 
-import android.app.Activity
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
-import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import kr.co.hanbit.activity.databinding.ActivityMainBinding
 
@@ -13,15 +11,13 @@ class MainActivity : AppCompatActivity() {
 
     val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
 
-    lateinit var resultListener: ActivityResultLauncher<Intent>
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
         // onActivityResult의 대안
-        resultListener = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
-            if (it.resultCode == Activity.RESULT_OK) {
+        val activityResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == RESULT_OK) {
                 val message = it.data?.getStringExtra("returnValue")
                 Toast.makeText(this, message, Toast.LENGTH_LONG).show()
             }
@@ -32,8 +28,6 @@ class MainActivity : AppCompatActivity() {
         intent.putExtra("from2", 2021)
 
         // startActivityForResult 의 대안
-        binding.btnStart.setOnClickListener {
-            resultListener.launch(intent)
-        }
+        binding.btnStart.setOnClickListener { activityResult.launch(intent) }
     }
 }
