@@ -1,58 +1,64 @@
 package com.example.cameraandgallery
 
 import android.Manifest
-import android.content.ContentValues
-import android.content.Intent
-import android.graphics.Bitmap
-import android.graphics.ImageDecoder
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.provider.MediaStore
+import android.os.Environment
 import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.FileProvider
 import com.example.cameraandgallery.databinding.ActivityMainBinding
-import java.io.IOException
-import java.text.SimpleDateFormat
+import java.io.File
 
+class MainActivity : AppCompatActivity() {
 
-class MainActivity : BaseActivity() {
+    var photoUri:Uri? = null
 
-    val PERM_STOREGE = 99
-    val PERM_CAMERA = 100
-    val REQ_CAMERA = 101
-    val REQ_STORAGE = 102
+    lateinit var cameraPermission:ActivityResultLauncher<String>
+    lateinit var storagePermission:ActivityResultLauncher<String>
+
+    lateinit var cameraLauncher:ActivityResultLauncher<Uri>
+    lateinit var galleryLauncher:ActivityResultLauncher<String>
 
     val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
-
-    var realUri: Uri? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
-        requirePermissions(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), PERM_STOREGE)
-    }
-
-    override fun permissionGranted(requestCode: Int) {
-        when(requestCode) {
-            PERM_STOREGE -> setViews()
-            PERM_CAMERA -> openCamera()
-        }
-    }
-
-    override fun permissionDenied(requestCode: Int) {
-        when(requestCode) {
-            PERM_STOREGE -> {
+        storagePermission = registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted ->
+            if(isGranted) {
+                setViews()
+            } else {
                 Toast.makeText(baseContext, "외부저장소 권한을 승인해야 앱을 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
                 finish()
             }
-            PERM_CAMERA -> Toast.makeText(baseContext, "카메라 권한을 승인해야 카메라를 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
         }
+
+        cameraPermission = registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted ->
+            if(isGranted) {
+                openCamera()
+            } else {
+                Toast.makeText(baseContext, "카메라 권한을 승인해야 카메라를 사용할 수 있습니다.", Toast.LENGTH_LONG).show()
+            }
+        }
+
+        cameraLauncher = registerForActivityResult(ActivityResultContracts.TakePicture()) { isSuccess  ->
+            if(isSuccess) { binding.imagePreview.setImageURI(photoUri) }
+        }
+
+        galleryLauncher = registerForActivityResult(ActivityResultContracts.GetContent()) { uri ->
+            binding.imagePreview.setImageURI(uri)
+        }
+
+        storagePermission.launch(Manifest.permission.WRITE_EXTERNAL_STORAGE)
     }
 
     fun setViews() {
         binding.buttonCamera.setOnClickListener {
-            requirePermissions(arrayOf(Manifest.permission.CAMERA), PERM_CAMERA)
+            cameraPermission.launch(Manifest.permission.CAMERA)
         }
         binding.buttonGallery.setOnClickListener {
             openGallery()
@@ -60,70 +66,22 @@ class MainActivity : BaseActivity() {
     }
 
     fun openCamera() {
-        val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+        val photoFile = File.createTempFile(
+            "IMG_",
+            ".jpg",
+            getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+        )
 
-        createImageUri(newFileName(), "image/jpg")?.let { uri ->
-            realUri = uri
-            intent.putExtra(MediaStore.EXTRA_OUTPUT, realUri)
-            startActivityForResult(intent, REQ_CAMERA)
-        }
+        photoUri = FileProvider.getUriForFile(
+            this,
+            "${packageName}.provider",
+            photoFile
+        )
+
+        cameraLauncher.launch(photoUri)
     }
 
     fun openGallery() {
-        val intent = Intent(Intent.ACTION_PICK)
-        intent.type = MediaStore.Images.Media.CONTENT_TYPE
-        startActivityForResult(intent, REQ_STORAGE)
-
-    }
-
-    fun newFileName() : String {
-        val sdf = SimpleDateFormat("yyyyMMdd_HHmmss")
-        val filename = sdf.format(System.currentTimeMillis())
-
-        return "$filename.jpg"
-    }
-
-    fun createImageUri(filename: String, mimeType: String) : Uri? {
-        var values = ContentValues()
-        values.put(MediaStore.Images.Media.DISPLAY_NAME, filename)
-        values.put(MediaStore.Images.Media.MIME_TYPE, mimeType)
-        return contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values)
-    }
-
-    fun loadBitmap(photoUri: Uri): Bitmap? {
-        var image: Bitmap? = null
-        try {
-            image = if (Build.VERSION.SDK_INT > 27) { // Api 버전별 이미지 처리
-                val source: ImageDecoder.Source =
-                        ImageDecoder.createSource(this.contentResolver, photoUri)
-                ImageDecoder.decodeBitmap(source)
-            } else {
-                MediaStore.Images.Media.getBitmap(this.contentResolver, photoUri)
-            }
-        } catch (e: IOException) {
-            e.printStackTrace()
-        }
-        return image
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == RESULT_OK) {
-            when (requestCode){
-                REQ_CAMERA -> {
-                    realUri?.let { uri ->
-                        val bitmap = loadBitmap(uri)
-                        binding.imagePreview.setImageBitmap(bitmap)
-
-                        realUri = null
-                    }
-                }
-                REQ_STORAGE -> {
-                    data?.data?.let { uri ->
-                        binding.imagePreview.setImageURI(uri)
-                    }
-                }
-            }
-        }
+        galleryLauncher.launch("image/*")
     }
 }
